FAT-889: Refactor the code
