+++ cypress/integration/users/delete-users.spec.js
@@ -8,6 +8,7 @@ import {
 } from '../../../interactors';
 
 import generateItemBarcode from '../../support/utils/generateItemBarcode';
+import EditRequest from "../../support/fragments/requests/edit-request";
 
 describe('Deleting user', () => {
   const lastName = 'Test-' + uuid();
@@ -102,7 +103,7 @@ describe('Deleting user', () => {
       .then(() => {
         verifyUserDeleteImpossible(userId);
 
-        cy.changeItemRequestApi({
+        EditRequest.updateRequestApi({
           ...Cypress.env('request'),
           status: 'Closed - Cancelled',
           cancelledByUserId: userId,
